+++ examples/vite-redux-toolkit-react-app/src/pages/HomePage.tsx
@@ -1,26 +1,20 @@
 import type React from 'react';
-import { useEffect, useState } from 'react';
+import { useMemo } from 'react';
 import { useStore } from 'react-redux';
 import {
   Box,
   CircularProgress,
 } from "@mui/material"
 import { getLayoutManagerInstance } from "@metacell/geppetto-meta-client/common/layout/LayoutManager";
-import '@metacell/geppetto-meta-ui/flex-layout/style/dark.scss'
+import '@metacell/geppetto-meta-client/common/layout/styles/dark.css'
+
 const HomePage = () => {
   const store = useStore();
-  const [LayoutComponent, setLayoutComponent] = useState<any | undefined>(undefined);
-  
-  useEffect(() => {
-    if (LayoutComponent === undefined) {
-      const myManager = getLayoutManagerInstance();
-      if (myManager) {
-        // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-        setLayoutComponent(myManager.getComponent() as React.ComponentType<any>);
-      }
-    }
-  }, [store, LayoutComponent])
-  
+  const LayoutComponent = useMemo(() => {
+    return getLayoutManagerInstance()?.getComponent()
+  }, [store])
+
+
   return (
     <Box sx={{
       display: 'flex',
@@ -30,7 +24,7 @@ const HomePage = () => {
       padding: 0,
       marginTop: 8
     }}>
-      {LayoutComponent === undefined ? <CircularProgress/> : <LayoutComponent/>}
+      {LayoutComponent === undefined ? <CircularProgress /> : <LayoutComponent />}
     </Box>
   );
 }